+++ extern/tiles.h
@@ -11,7 +11,7 @@
 #define MaxLONGINT 2147483647
 
 void tiles(
-    int the_tiles[],           // provided array contains returned tiles (tile indices)
+    size_t the_tiles[],           // provided array contains returned tiles (tile indices)
     int num_tilings,           // number of tile indices to be returned in tiles
     int memory_size,           // total number of possible tiles
     float floats[],            // array of floating point variables
